Restructuring into single package
