+++ src/CyberCity/API/TrafficLight.py
@@ -17,19 +17,19 @@ class TrafficLight():
         Initialize a Traffic Light Object
         Args:
             name (str): Name of the Traffic Light
-            startCoil (int): the coil for GREEN Light, YELLOW is +1, RED is +2
+            startCoil (int): Modbus coil address for GREEN Light, YELLOW is +1, RED is +2
         """
         self.name: str = name
         """ Name of the Traffic Light """
         self.startCoil: int = startCoil
-        """ The coil for GREEN Light, YELLOW is +1, RED is +2 """
+        """ Modbus coil address for GREEN Light, YELLOW is +1, RED is +2 """
 
         self.RED = False
-        """ RED Light """
+        """ RED Light state """
         self.YELLOW = False
-        """ YELLOW Light """
+        """ YELLOW Light state """
         self.GREEN = False
-        """ GREEN Light """
+        """ GREEN Light state """
 
     def __str__(self) -> str:
         """
@@ -137,110 +137,5 @@ class TrafficLight():
             result = self.client.read_coils(self.startCoil, 3)
             self.client.close()
             return result.bits[:3]
-        else:
-            return
-
-class Power():
-    HOST = '127.0.0.1'
-    PORT = 502
-    client = ModbusClient(HOST, PORT)
-
-    mainGrid = True
-
-    @staticmethod
-    def mainGridOff():
-        Power.mainGrid = False
-
-    @staticmethod
-    def mainGridOn():
-        Power.mainGrid = True
-
-    def __init__(self, name: str, globalPowerCoil: int):
-        self.name = name
-        self.globalPowerCoil = globalPowerCoil
-
-        self.globalPower = True
-
-    def __str__(self):
-        return f'{"⚡" if self.globalPower else "⚫"}'
-
-    def globalPowerOff(self):
-        self.globalPower = False
-
-    def globalPowerOn(self):
-        self.globalPower = True
-
-    def getGlobalPower(self) -> bool:
-        return self.globalPower
-
-    def write(self):
-        if Power.client.connect():
-            if Power.mainGrid:
-                self.client.write_coil(self.globalPowerCoil, self.globalPower)
-            else:
-                self.client.write_coil(self.globalPowerCoil, False)
-            self.client.close()
-        else:
-            return
-
-    def read(self) -> bool:
-        if Power.client.connect():
-            result = self.client.read_coils(self.globalPowerCoil, 1)
-            self.client.close()
-            return result.bits[0]
-        else:
-            return
-
-class District():
-    def __init__(self, name: str, globalPowerCoil: int):
-        self.name = name
-
-        self.globalPower = Power(self.name, globalPowerCoil)
-
-    def __str__(self):
-        return f'{self.name} => {self.globalPower} '
-
-class SpecialDistrict(District):
-    def __init__(self, name: str, globalPowerCoil: int, host: str, port: int, startCoil: int = 0):
-        super().__init__(name, globalPowerCoil)
-
-        self.HOST = host
-        self.PORT = port
-        self.CLIENT = ModbusClient(self.HOST, self.PORT)
-
-        self.localPower = True
-        self.localGenerator = False
-
-        self.startCoil = startCoil
-
-    def __str__(self):
-        return super().__str__() + f'{"🔌" if self.localPower else "⚫"} {"🔋" if self.localGenerator else "⚫"}'
-
-    def localPowerOff(self):
-        self.localPower = False
-    def localPowerOn(self):
-        self.localPower = True
-    def getLocalPower(self) -> bool:
-        return self.localPower
-
-    def localGeneratorOff(self):
-        self.localGenerator = False
-    def localGeneratorOn(self):
-        self.localGenerator = True
-    def getLocalGenerator(self) -> bool:
-        return self.localGenerator
-
-    def write(self):
-        if self.CLIENT.connect():
-            self.CLIENT.write_coils(self.startCoil, [self.localPower, self.localGenerator])
-            self.CLIENT.close()
-        else:
-            return
-
-    def read(self) -> List[bool]:
-        if self.CLIENT.connect():
-            result = self.CLIENT.read_coils(self.startCoil, 2)
-            self.CLIENT.close()
-            return result.bits[:2]
         else:
             return
